Don't rebuild wifi-ap snap when already build locally

+++ tests/lib/prepare-all.sh
@@ -6,6 +6,13 @@ if [ -n "$SNAP_CHANNEL" ] ; then
 	exit 0
 fi
 
+# If there is a wifi-ap snap prebuilt for us, lets take
+# that one to speed things up.
+if [ -e /home/wifi-ap/wifi-ap_*_amd64.snap ] ; then
+	exit 0
+fi
+
+
 # Setup classic snap and build the wifi-ap snap in there
 snap install --devmode --beta classic
 cat <<-EOF > /home/test/build-snap.sh
